refactors Reconcile to support Update and prepare for more resource types
Puts "ensure CephCluster" code in its own function with the ability to add more
similar functions that ensure the existence of other resources.

Updates a CephCluster resource if its Spec does not match the desired state.

Stopped accessing r.reqLogger from Reconcile since it's possible for two
Reconciles to run concurrently. We set the name and namespace of the resource
being reconciled on the logger as attributes, so we don't want two Reconciles
fighting over that.