adds OCSInitialData resource
This resource represents the initial data that should be created when the
operator gets installed. Currently it includes two demo StorageClass resources,
which should be changed to match what OCS wants to install in production.

Deleting the OCSInitialData resource will cause it to be auto-recreated, and
then all initial data will be reset or recreated as necessary to restore their
initial state.