Move manifests to different location in registry image
The default WORKDIR for the base image
(quay.io/openshift/origin-operator-registry) is `/registry`. The
database intializater was being run with this directory as the manifests
directory, and it wrote the generate database file in the default
WORKDIR, ie. `registry`.

The intializer walks the manifests directory and tries to add all files
it finds into the database, which includes the generated database file.
It fails to parse the database file and fails. In an older version of
the base image, the failure to parse was not an error. An update to the
base image, has made the failure to parse an error. This is causing the
building of the ocs-registry image to fail.

This patch moves the copied manifests to another directory, and updates
the intializer command to use the new directory. With this, the
ocs-registry image builds successfully.

Fixes #17.

Signed-off-by: Kaushal M <kshlmster@gmail.com>