Hide Rook/NooBaa CRDs in operator catalog
This commit removes the Rook and NooBaa CRDs from the owned CRD list in
the CSV. The manifest files for the Rook and NooBaa CRDs are still
present in the catalog directory. Also, the operator deployments for the
rook-ceph and noobaa-operator are still present.

The generated catalog bundle, bundles all the CRD manifests from the
catlog directory, even if they are not present in the CSV. When the OLM
deploys the ocs-operator CSV from this catalog, it creates all the
included CRDs in the cluster, and starts all the operators.

The end result of this, is that only the OCS operator CRDs are available
in the Console, with the other CRDs being hidden. Rook and NooBaa CRDs
not being available from the Console is a good thing for OCS, as we do
not want users to be manually creating their own CephClusters or other
Rook/NooBaa resources.

Users can create a StorageCluster resource from the UI, which leads to
the ocs-opertor creating a CephCluster resource, and the rook-ceph
operator creating the actual ceph cluster. All of this still works, as
the CRDs required are present in the cluster.