api: Add Portable property to StorageDeviceSet
StorageDeviceSet.Portable is translated to
StorageClassDeviceSet.Portable by the ocs-operator.

Rook's StorageClassDeviceSet introduced a new property, 'Portable' which
defines if a OSDs are portable. If false (default) Rook pins OSDs to
specific hosts. In a cloud environment like AWS, the OCS operator will
need to set this property to ensure that the OSDs get scheduled
correctly.

Signed-off-by: Kaushal M <kshlmster@gmail.com>