added openAPI validation to StorageCluster CRD
Currently, .storageDeviceSet.Count allows negative value.
This should not be allowed as it will render the cluster
unusuable when large negative values are set.

Say for a default cluster with .storageDeviceSet.Count=3,
if user updates the CR with .storageDeviceSet.Count=-6
this is not a valid state that can be achieved.

With this validation, CR creation will fail if negative
count value is provided.

Signed-off-by: Umanga Chapagain <chapagainumanga@gmail.com>