storagecluster: Correctly set tolerations for StorageClassDeviceSets
Tolerations were being set on copies of items returned by range, instead
of the actual items themselves. Becuase of this, the CephCluster
resource created by the StorageCluster controller did not have
tolerations set for the StorageClassDeviceSets.

This commit changes the loop used to set the tolerations to use slice
index, to actually modify the object in slice.

Signed-off-by: Kaushal M <kshlmster@gmail.com>