storagecluster: Do not set podAntiAffinity for device sets
Commit 062137e (storagecluster: Correctly set tolerations for
StorageClassDeviceSets) ensures that placement configuration is
correctly set for StorageClassDeviceSets, and passed along to Rook.

With this, podAntiAffinity is being set for StorageClassDeviceSets. The
set podAntiAffinity being set, does not have a topologyKey set, which
causes failure of the OSD prepare jobs to be scheduled and prevents
complete deployment of the storage cluster.

This commit removes the podAntiAffinity on StorageClassDeviceSets, to
allow pods to be scheduled.

Signed-off-by: Kaushal M <kshlmster@gmail.com>