Adding upgrade-test script
This script deploys the OCS cluster using a working registry image in quay.io.
It then upgrades the OCS cluster to the latest version of image or the
latest build in openshift-ci. The idea is to start with a previous
version that we know is working, and make sure the cluster can be
upgraded to the latest build. The actual image creation is done outside
of the script.