Add option to install rook/ceph tools pod during ocs-ci functest run
The tools pod is no longer installed automatically by ocs-operator. That
responsibility moved to the ocs-ci test suite. Unfortunately since we
don't install the ocs-operator using the ocs-ci test suite, but merely
point ocs-ci at a pre-installed ocs-operator deployment, the tools pod
is not installed. This causes ocs-ci to fail.

To work around this temporarily, an environment variable is used to
toggle installing the tools pod when ocs-ci is being run.

Signed-off-by: David Vossel <davidvossel@gmail.com>