Add zone detection to StorageCluster reconciliation
This commit adds the ability to dynamically determine a best-effort
configuration for both the Ceph failure domain and the OSD Pod
distribution for a StorageCluster. To start, we look for the values of
all topology labels of all nodes in the storage cluster. If three or
more "zones" are detected across all nodes, the Ceph failure domain is
set to "zone" and the StorageClassDeviceSets are spread across three
zones and set to portable. If fewer than three zones are detected, the
Ceph failure domain is set to "host" and the StorageClassDeviceSets are
marked as not portable.

NOTE: This changes the behavior of the operator's generation of
Rook-Ceph StorageClassDeviceSets. Previously, a number of
StorageClassDeviceSets were created equal to a given StorageDeviceSet's
Count. Now, the number of StorageClassDeviceSets created is equal to the
value of a given StorageDeviceSet's Replica. StorageDeviceSet.Count is
now mapped to StorageClassDeviceSet.Count.

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>