Use "rack" as default failure domain
When creating a StorageCluster on nodes across fewer than three zones,
the OSD Pods cannot be guaranteed to distribute evenly across the
storage nodes. This change introduces the notion of "racks" as the
default failure domain rather than "host", and works to sort the storage
nodes into three or more virtual groups. These groups are called "racks"
to allow for compatibility with the default Ceph CRUSH hierarchy.
Topology labels for racks can also be pre-defined by the admin, and will
be recognized by the operator.

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>