Bug 1771844: storagecluster: Use correct CephCluster name when ...
... waiting for CephCluster to create NooBaa. Without this, the
contoller will not create a NooBaa resource.

New generator functions have been added to create the prefixed
CephCluster name, and these functions are used wherever the name of the
CephCluster is required.

The unit tests also have been updated to use new mock objects for
CephCluster created with the above functions.

Backport of #290.

Signed-off-by: Kaushal M <kshlmster@gmail.com>