Bug 1769596: Fixes panic of ocs-operator when running vmware
This commit fixes the panic of the ocs-operator when running in an vmware based ocp cluster. The operator crashes because its not able to assign a valid rack to the nodes as nodes in vmware based ocp cluster doesn't have any valid topology keys (https://github.com/openshift/ocs-operator/blob/master/pkg/controller/storagecluster/reconcile.go#L46-L50).So the current work around is to assign a rack with min number of nodes to the new node if the operator is not able to find a valid rack based on the topology key checks(https://github.com/openshift/ocs-operator/blob/master/pkg/controller/storagecluster/reconcile.go#L498-L530).

Signed-off-by: Ashish Ranjan <ashishranjan738@gmail.com>