storagecluster: Do not set ownerRefs for StorageClasses
The StorageClass resources are non-namespaced resources, and had their
ownerRefs set to the StorageCluster resource which is a namespaced
resource in the openshift-storage namespace. Having this sort of
cross-namespace ownerRefs, could cause the other children of the
StorageCluster to be inadverdently deleted by the GC. Ref [1][1].

Removing the ownerRefs on the StorageClass objects avoids this.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1691546

Signed-off-by: Kaushal M <kshlmster@gmail.com>