sets mon_osd_full_ratio=.85 & mon_osd_nearfull_ratio =.75
mon_osd_full_ratio is percentage of disk space used before an OSD
is considered full and defaults to .95. mon_osd_nearfull_ratio is
percentage of disk space used before an OSD is considered nearfull
and defaults to .85. These defaults are not optimal for OCS. So,
overriding the defaults with optimal values.

Signed-off-by: Umanga Chapagain <chapagainumanga@gmail.com>