Adds new OLM annotations to CSV
Adding 2 new recommended annotations to OCS CSV.

"operatorframework.io/suggested-namespace" is used to
suggest namespace for the OCS Operator to be installed in.
Admin is still allowed to select other namespace.

"operatorframework.io/cluster-monitoring=true" is used to
inform admin that monitoring is supported by the operator
and provide them with a checkbox to opt-in, following
which the namespace will be labeled appropriately.
Admin can opt-out but will be unable to use Storage
Dashboard & other OCS monitoring & alerting features.

Signed-off-by: Umanga Chapagain <chapagainumanga@gmail.com>