Removed hardcoded portable=true
The OCP console used to require us to hardcode portable=true.
Now it explicitly sets portable=true so there is no longer need to
hardcode it in the operator code. Removing the hardcoded value from
the operator code also makes it possible to set the value of portable
without setting a placement as a workaround.

Signed-off-by: rohan47 <rohgupta@redhat.com>