Add ROOK_CSI_ALLOW_UNSUPPORTED_VERSION to rook operator env
The ceph-CSI version of both upstream and downstream are different
ex: the upstream versions are 2.0.0,2.1.0 etc but downstream
built version are release-4.3,release4.4 etc.
we cannot verify the supported downstream version with upstream
version.so setting ROOK_CSI_ALLOW_UNSUPPORTED_VERSION to "true"
in rook operator ENV to skip the version check for the
CSI pods.

Signed-off-by: Madhu Rajanna <madhupr007@gmail.com>