adds support for appregistry format
This adds support for bulilding ocs-registry in appregistry
format in addition to the bundle format that we already have.
It gives us flexibilty to choose between formats and experiment
with it.

`make ocs-registry` will result in appregistry format image.
This is done by temporarily converting the directory structure
expected by bundle format to the one expected by appregistry
under build/_output/appregistry directory.

Signed-off-by: Umanga Chapagain <chapagainumanga@gmail.com>