adds structural schema to CRDs
structural schema is mandatory on new versions of Kubernetes
and OpenShift without which CRD validation fails. This is
different from OpenAPI schema validation and is also used
by OpenShift Console to display information about the Spec
to the users.

`make update-generated` adds structural schema to CRD. Internally
it calls update-codegen-crds which wraps controller-gen for
updating CRDs.

Signed-off-by: Umanga Chapagain <chapagainumanga@gmail.com>