fix: add missing Rook-Ceph operator env variable
For the OBC code to work properly, we need the Rook-Ceph operator to run
with `ROOK_OBC_WATCH_OPERATOR_NAMESPACE` set to `true` so that each the
OBC will look for requests in the cluster namespace. This allows us to
run multiple OBCs in different namespaces with a single operator.
Even though this is not something the OCS operator supports, it is
still required for Rook-Ceph to operate normally.

Signed-off-by: Sébastien Han <seb@redhat.com>