Propagate the cleanup policy
At the time of deletion of the StorageCluster, we want to check the
cleanup policy and propagate the same to the objects for which
StorageCluster has owner reference.

This commit applies the cleanup policy for rook. The label that
ocs-operator refers to is "cleanup.ocs.openshift.io" and the only valid
value at the time of this patch is "yes-really-destroy-data".

Updated the references to rook to get the commits that have the cleanup
policy definition.

Signed-off-by: Raghavendra Talur <raghavendra.talur@gmail.com>