independent mode: expose rgw endpoint as a label
When doing indepenent mode, we must instruct Noobaa where to find the
external Ceph rgw endpointhen we fetch the externalClusterDetailsSecret
Secret, we extract the Rgw StorageClass as well as its parameters.
Currently, we only have a single parameter called "endpoint", which is
used to configure OBC.
Now we also expose that endpoint as a label on the Nooba CR so Noobaa
can determine whether the endpoint is external or not.

Signed-off-by: Sébastien Han <seb@redhat.com>
(cherry picked from commit 91a38cb91be3a26218e25e2ccc6cc997c63cb270)