encore rgw endpoint for external mode
Since labels value cannot contain any `:` when Noobaa reads it when an
error like:

```
Error while reconciling: NooBaa.noobaa.io "noobaa" is invalid:
metadata.labels: Invalid value: "10.70.56.202:8080":
a valid label must be an empty string or consist of alphanumeric characters, '-', '_' or '.',
and must start and end with an alphanumeric character
(e.g. 'MyValue', or 'my_value', or '12345', regex used for validation is '(([A-Za-z0-9][-A-Za-z0-9_.]*)?[A-Za-z0-9])?')
```

So let's encode the endpoint and decode it later with Nooba.

Signed-off-by: Sébastien Han <seb@redhat.com>
(cherry picked from commit d27a046c017cbfc06c4b16c6c79964cce2221499)