Remove the host from the crush map
This change will remove the host associated with the failed osd from the
crush map. When node removal process is done and all OSDs have been
removed from that host, the empty host will be removed from crush map.
The user will be notified that the error while removing host should be
ignored if there are other OSDs present on the same host

Signed-off-by: Servesha Dudhgaonkar <sdudhgao@redhat.com>