purge-osd-script: undo indentation change of PR #563
PR #563 accidentally removed indentation in the osd purge script
when it intended to only remove trailing white spaces. It also
by accident mixed this with a different commit. This makes backporting
of for instance #589 unnecessarily difficult. As a first step, this
patch reverts the whitespacing to the state before #563, and the next
patch will introduce only the trailing whitespace cleanup.

Signed-off-by: Michael Adam <obnox@redhat.com>