osd-purge-script: remove trailing white spaces
This is the whitespace cleanup that @rajatsingh25aug intended to do in
PR #563. But indenting white spaces were removed too, and the commits
were a bit messed up.

Committing this separately to ease backporting of future patches on top
of backport of PR #589. If we ever need to backport another patch in
this area, we could just pick this patch first and then continue with
clean cherry-picks.

Signed-off-by: Michael Adam <obnox@redhat.com>