Fix segfault with empty labelselector in newStorageClassDeviceSets
getPlacements only properly initalized placement.NodeAffinity if the
label selector was nil or contained terms. If the label selector is
specified but empty, the intended way to avoid label selectors entirely,
then it would not be initialized, and newStorageClassDeviceSets would
crash on nil pointer dereference.

This patch fixes this by adding proper initialization to getPlacement().

Signed-off-by: Michael Adam <obnox@redhat.com>