Add containerized build
This adds the option to do a containerized build of the ocs-operator so
that developers can be sure they are compiling in a shared environment.
The current "build" target is abstracted into "build-go" in anticipation
for changing the OpenShift ci-operator integration to use it as the
binary build command, at which point "build" can be pointed to
"build-container" so all developers can run it by default.

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>