Remove build from ocs-operator-ci
The build step is unneccessary, as the previous targets will fail if the
code has any compile errors. In addition, this is in anticipation of
migrating the "build" target to "build-container", at which point we
wouldn't be able to run the "build" target in the OpenShift ci-operator
(since it runs the "ocs-operator-ci" target in a container already).

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>