Create only those StorageClasses provided in the external JSON input file
In External Cluster mode, create only those StorageClasses which are
provided in JSON input file.

Added 'setRookCSICephFS' function, which could toggle Rook CephFS CSI driver,
which will enable or disable (the driver) according to whether CephFS
storageclass is available (or not) in the external JSON input.

Corrected the tests to work with latest changes.

Signed-off-by: Arun Kumar Mohan <amohan@redhat.com>