StorageCluster: propagate rook uninstall mode
Rook waits for the deletion of the PVCs before deleting the cephCluster.
To override the wait, we can set the "allowUninstallWithVolumes" option
in the Cleanup Policy of the cephCluster.

With this commit, if the UninstallMode is set to forced in the
StorageCluster, we set the allowUninstallWithVolumes option to true in
the cephCluster.

Co-authored-by: Nitin Goyal <nigoyal@redhat.com>
Signed-off-by: Raghavendra Talur <raghavendra.talur@gmail.com>