Add TargetSizeRatio of .49 to RGW data pool
The purpose of the `target_size_ratio` is to configure the pool for some
pre-optimization of pg placement when the autoscaler is active. If the
`target_size_ratio` is set, then up to that threshold of capacity, the
pre-optimization will be active. Above the threshold, it is changing the normal
non-optimized mode.

This change sets the `target_size_ratio` for the RGW pool to `.49` like for the
rbd pools and the cephfs data pool as recommended by ceph experts (starting
with RHCS 4.1).

Signed-off-by: Michael Adam <obnox@redhat.com>