storagecluster: delete cephCluster before ceph resources
rook ceph requires that the deletion request for the cephCluster should
come before the delete request for the other ceph resources. Not doing
so can result in some of the ceph resources being stuck in deletion
phase forever.

Deleting the cephCluster first also helps in the cases where the
CephCluster has failed to install.

Signed-off-by: Raghavendra Talur <raghavendra.talur@gmail.com>