Update operator capabilities to level 4
This commit updates the OCS Operator's capabilities level to 4, "Deep
Insights". Per the Operator SDK documentation[1], we now fulfill the
following requirements:

- Operator exposing metrics about its health
- Operator exposes health and performance metrics about the Operand
- Operand sends useful alerts
- Custom Resources emit custom events

The following criteria was determined by the Operator SDK team to not be
required, and indeed may soon be removed:

- Operator leverages Operator Metering

Though we may not fully have metrics for all aspects of our operations,
simply having the level we currently do is sufficient to meet the
criteria.

[1] https://sdk.operatorframework.io/docs/advanced-topics/operator-capabilities/operator-capabilities/#level-4---deep-insights

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>