Move Snapshot management strategies into Ceph management
Prior to this change, Snapshot management strategies were
in one monolithic struct under the ManagedResourcesSpec. This
would allow users to fall out of sync between their Snapshot
management policies and the management policies of their
CephFilesystem and CephBlockPool. This patch enables users to
disable StorageClass management as part of the ReconcileStrategy
of each Ceph storage type.

Signed-off-by: Elise Gafford <egafford@redhat.com>