pkg: add arbiter spec to the Storage Cluster Spec
This commit adds the arbiter mode of deployment to the Storage Cluster
spec. Two new fields are added to the Storage Cluster

* ArbiterSpec
* NodeTopologies

If ArbiterSpec.Enable is set to true *and* if
NodeTopologies.ArbiterLocation is set then the Ceph Cluster is
configured with arbiter enabled.

If ArbiterSpec.Enable is true and no NodeTopologies.ArbiterLocation is
provided then the Storage Cluster fails to reconcile.

If ArbiterSpec.Enable is false then NodeTopologies.ArbiterLocation is
ignored even if set.

The arbiter mode differs from the non-arbiter mode in the following
ways:

* The minimum and the maximum number of MONs are 5 instead of 3.
* The OSDs run only in 2 of the 3 failure domains. The third failure
  domain is only used to host a MON pod.
* Mininum number of OSDs in each failure domain is 2 instead of 1.
* The CRUSH Map is configured in replica 4 configuration instead of 3.
* The OSDs can be scaled in multiples of 2 and not 3.
* The placement algorithm doesn't restrict the arbiter MON pod to nodes
  which have the default/user-provided OCS node label. Rest of the ceph
  pods have the nodeAffinity same as in the non-arbiter mode.
* The arbiter MON pod is scheduled to run on any of the nodes in the
  ArbiterLocation in the failure domain.
* The arbiter MON pod has toleration for the master node taint by default. It
  can be disabled by setting ArbiterSpec.DisableMasterNodeToleration to true.

ocs-operator's placement strategy for non-arbiter mode is to select a failure
domain and then spread the ceph resources across the failure domain.

Keeping with the same principle, ocs-operator determines the failure domain
automatically in the the arbiter mode too and tries to match one of the failure
domain locations to the NodeTopologies.ArbiterLocation.

In both the modes, it might sometimes be desirable that the user gets to
specify the failure domain through the Storage Cluster and not rely on the
automatic detection. They may also wish to specify the node topology
themselves.  For this reason, we did not add the ArbiterLocation field to the
ArbiterSpec and added it to the TopologyMap struct instead.

Work needs to be done to reconcile StorageClusterSpec.FlexibleScaling,
StorageClusterSpec.NodeTopologies.FailureDomain, and
StorageClusterStatus.FailureDomain to let users choose the node topology and
the failure domain. This can be done in the future, we have made only those
changes that are necessary to not break the API later.

Signed-off-by: Raghavendra Talur <raghavendra.talur@gmail.com>