cephcluster: refactor `throttleStorageDevices`: separate data from code
Instead of having a list of known fast devices, and a list of known slow
devices, and hard-coding in the throttleStorageDevices function for which
provisioner each list is applicable, create known disk type that
encapsulates the provisioner, storage class name and whether this is
known to be fast or slow. The function then just becomes a simple loop.

IMHO, this makes the code much more obvious and easier to extend in the
future.

Signed-off-by: Michael Adam <obnox@redhat.com>