storageclasses: consolidate checks for skipping obc storage class
In principle, newStorageClasse() provides the in-memory storage
class structures, and createStorageClasses() creates them on the
cluster using the client.

Now we have some mechanisms to select which storage classes to
create and which ones to skip.  Up to now, this logic was scattered
across the two functions: The newStorageClasses() pre-selected whether
to add the OBC storage class under certain circumstances, and recently,
the configuration for reconcile strategies was added to the storage
cluster, and the filtering based on this is present in
createStorageClasses().

This patch consolidates the checks in the createStorageClasses function,
making the newStorageClasses function very simple again.

One small side effect in behavior is that previously, any possible error
from GetPlatform() was ignored, and we were *not* creating the OBC
storage class in this case. Now we correctly error out if GetPlatform
fails.

Signed-off-by: Michael Adam <obnox@redhat.com>