Enable non-default StorageClusters for unit testing
Prior to this change, creation of a default StorageCluster was
tightly coupled with test initialization functions. As such, many
of our unit tests only cover the default happy path. This change
introduces the option of providing a non-default StorageCluster
for test initialization while leaving the signatures of existing
initialization functions unchanged.

Signed-off-by: egafford <egafford@redhat.com>