Adjust mon_max_pg_per_osd for Rook-Ceph
This is a temporary workaround for a race condition with the Ceph PG
autoscaler. In cases where the autoscaler only detects our default RBD
and CephFS pools it will set the PG count for them to 128 instead of the
usual 32. In small clusters, this prevents the creation of additional
pools, e.g. for RGW. Setting mon_max_pg_per_osd to a higher value helps
account for the temporary state where few pools ar eusing a high number
of PGs.

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>