Use Golang 1.15	and fix test dependencies
Golang 1.15 has security fixes to Golang std library.

Currently, `go install` is unable to install ginkgo which
causes build failures during E2E tests. The workaround is
to install ginkgo and gomega using `go get` and turn off
the modules so that it doesn't update go.mod and go.sum files.

Once we start using Go 1.16 we can switch back to `go install`
as it is improved and recommended way to install dependencies.

Signed-off-by: Umanga Chapagain <chapagainumanga@gmail.com>