StorageCluster: fix osd placement logic for the OSDs
The logic to determine the location affinity for the OSDs should be
based on the minimum DeviceSetReplica for a Storage Cluster
configuration and not on the replica count of the StorageDeviceSet.

TODO: The function to generate the StorageClassDevicesSets should be
refactored further. This is a short term fix to make it easy to
backport to other branches.

Signed-off-by: Raghavendra Talur <raghavendra.talur@gmail.com>