tests: use fake objects provided by the fake client
The fake client stores the objects after adding a resource version to
them. This is a breaking change introduced in
https://github.com/kubernetes-sigs/controller-runtime/pull/1306.
Therefore we cannot use the fake object that we provided as input to the
the fake client and should use the object obtained from the Get
operation.

The helper function returns the objects after obtaining them from the
fake client and therefore these objects cannot be used for create
functions.

Signed-off-by: Raghavendra Talur <raghavendra.talur@gmail.com>