fix merge strategy for newCephDaemonResources
newCephDaemonResources was taking custom resource requirements
for multiple daemons but only returning resource requirements
of ceph-mgr and mon.

This patch changes it to take all custom resource requirements
and merge it with the defaults. If key exists, it's value will
be overwritten and if it doesn't exist, it will be appended.

Signed-off-by: Umanga Chapagain <chapagainumanga@gmail.com>