Added Operator-Conditions to report correct csv status
Set() sets the specific condition on the operator's
OperatorCondition to the provided status. If the condition is not
present, it is added to the CR.
Currently, the only supported condition is "Upradeable".
Set() is a part of Condition interface.

NewCondition() returns a new Condition interface using the provided client
for the specified conditionType,
which is called in NewUpgradeable() in controllers/storagecluster/operator_conditions.go

Signed off by: Priyanka Jiandani <pjiandan@redhat.com>