csv: hide OCS CSV from UI
This commit adds a special annotation to the CSV to mark it as a
"non-standalone" operator. This annotation is exclusively used by the
OpenShift Console as a way to omit the given CSV/bundle from displaying
in the Operator Hub as well as the list of Installed Operators.

The only current exception to this feature is when the annotated CSV
reports an error. In that scenario it will show up in the Installed
Operators list until the problem is resolved and the CSV reports it's
installed correctly.

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>