Update builder Dockerfile for OpenShift CI usage
Update build/Dockerfile.build to still allow local building and also
allow for direct usage by OpenShift CI. build/Dockerfile.build will
become the de-facto image used for building ocs-operator in OpenShift
CI.

Because it is the image used for building in OpenShift CI, it should
*not* implicitly update dependencies as part of the build. All updates
to Go module dependencies should be tracked in the source code so that
builds are reproducible.

CI tests were updated to enforce that the Go module deps are updated in
addition to other generated files.

Signed-off-by: Blaine Gardner <blaine.gardner@redhat.com>