storagecluster: move topology map reconciliation into its own resourceManager
When we enabled a NooBaa-standalone StorageCluster, we were still trying
to reconcile the topology map, even though in such a deployment there
would be no need for any nodes to be labeled for OCS/Ceph.

The fix was fairly straightforward, but rather than adding another
conditional to the existing `if` statement it's a slightly tidier
implementation to create a new resourceManager for reconciling the
topology map. This lets us take advantage of the existing conditionals
for the different types of StorageClusters as well as the standard error
handling we do when we loop through the full reconciliation steps.

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>