controllers: always fetch the secret and store it in the map
after restarting a operator it was getting returned if there is no
change in the secret without saving the secret into the map and
cephcluster func is complaining that it can not found the secret in the
map. change the order of the if statement to always save the secret data
in the map.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>