Update source-manifests.sh
This makes the script use oc tool to extract the necessary files -- CSVs
and CRDs from rook and noobaa images. This should be less prone to
errors as it does not try to run the containers. It should also help us
sync with downstream where we can't use podman/docker to run the
containers.

Signed-off-by: Boris Ranto <branto@redhat.com>